chore(infrastructure): Refactor diff base parsing into its own class
