+++ test/screenshot/infra/lib/constants.js
@@ -61,15 +61,19 @@ module.exports = {
 
   ExitCode: {
     OK: 0,
-    UNKNOWN_ERROR: 11,
-    SIGINT: 12, // ctrl-c
-    SIGTERM: 13, // kill
-    UNSUPPORTED_CLI_COMMAND: 14,
-    HTTP_PORT_ALREADY_IN_USE: 15,
-    MISSING_ENV_VAR: 16,
-    UNSUPPORTED_EXTERNAL_PR: 17,
-    UNHANDLED_PROMISE_REJECTION: 18,
-    UNCAUGHT_EXCEPTION: 19,
-    CHANGES_FOUND: 20,
+
+    /** ctrl-c */
+    SIGINT: 11,
+
+    /** kill */
+    SIGTERM: 12,
+
+    UNKNOWN_ERROR: 13,
+    UNCAUGHT_EXCEPTION: 14,
+    UNHANDLED_PROMISE_REJECTION: 15,
+    UNSUPPORTED_CLI_COMMAND: 16,
+    MISSING_ENV_VAR: 17,
+    HTTP_PORT_ALREADY_IN_USE: 18,
+    CHANGES_FOUND: 19,
   },
 };
