+++ test/screenshot/infra/lib/controller.js
@@ -100,7 +100,7 @@ class Controller {
    * @return {!Promise<!mdc.proto.ReportData>}
    */
   async initForCapture() {
-    const isOnline = await this.cli_.isOnline();
+    const isOnline = this.cli_.isOnline();
     const shouldFetch = this.cli_.shouldFetch;
     if (isOnline && shouldFetch) {
       await this.gitRepo_.fetch();
@@ -115,7 +115,7 @@ class Controller {
    * @return {!Promise<!mdc.proto.ReportData>}
    */
   async initForDemo() {
-    const isOnline = await this.cli_.isOnline();
+    const isOnline = this.cli_.isOnline();
     const shouldFetch = this.cli_.shouldFetch;
     if (isOnline && shouldFetch) {
       await this.gitRepo_.fetch();
@@ -207,11 +207,12 @@ class Controller {
     const boldRed = Logger.colors.bold.red;
     const boldGreen = Logger.colors.bold.green;
 
+    this.logger_.log('\n');
     if (numChanges > 0) {
-      this.logger_.error(boldRed(`\n\n${numChanges} screenshot${numChanges === 1 ? '' : 's'} changed!\n`));
+      this.logger_.error(boldRed(`${numChanges} screenshot${numChanges === 1 ? '' : 's'} changed!\n`));
       this.logger_.log('Diff report:', boldRed(reportData.meta.report_html_file.public_url));
     } else {
-      this.logger_.log(boldGreen(`\n\n${numChanges} screenshot${numChanges === 1 ? '' : 's'} changed!\n`));
+      this.logger_.log(boldGreen('0 screenshots changed!\n'));
       this.logger_.log('Diff report:', boldGreen(reportData.meta.report_html_file.public_url));
     }
 
@@ -235,7 +236,11 @@ class Controller {
       reportData.screenshots.added_screenshot_list.length +
       reportData.screenshots.removed_screenshot_list.length;
 
-    return numChanges > 0 ? ExitCode.CHANGES_FOUND : ExitCode.OK;
+    const isOnline = this.cli_.isOnline();
+    if (isOnline && numChanges > 0) {
+      return ExitCode.CHANGES_FOUND;
+    }
+    return ExitCode.OK;
   }
 
   /**
