+++ test/screenshot/infra/lib/report-builder.js
@@ -30,7 +30,9 @@ const {Approvals, DiffImageResult, Dimensions, GitRevision, GitStatus, GoldenScr
 const {ReportData, ReportMeta, Screenshot, Screenshots, ScreenshotList, TestFile, User, UserAgents} = mdcProto;
 const {InclusionType, CaptureState} = Screenshot;
 
+const CbtApi = require('./cbt-api');
 const Cli = require('./cli');
+const DiffBaseParser = require('./diff-base-parser');
 const FileCache = require('./file-cache');
 const GitHubApi = require('./github-api');
 const GitRepo = require('./git-repo');
@@ -46,10 +48,22 @@ const TEMP_DIR = os.tmpdir();
 class ReportBuilder {
   constructor() {
     /**
+     * @type {!CbtApi}
+     * @private
+     */
+    this.cbtApi_ = new CbtApi();
+
+    /**
      * @type {!Cli}
      * @private
      */
     this.cli_ = new Cli();
+
+    /**
+     * @type {!DiffBaseParser}
+     * @private
+     */
+    this.diffBaseParser_ = new DiffBaseParser();
 
     /**
      * @type {!FileCache}
@@ -115,8 +129,8 @@ class ReportBuilder {
   async initForCapture() {
     this.logger_.foldStart('screenshot.init', 'ReportBuilder#initForCapture()');
 
-    /** @type {boolean} */
-    const isOnline = await this.cli_.isOnline();
+    await this.cbtApi_.fetchAvailableDevices();
+
     /** @type {!mdc.proto.ReportMeta} */
     const reportMeta = await this.createReportMetaProto_();
     /** @type {!mdc.proto.UserAgents} */
@@ -125,7 +139,7 @@ class ReportBuilder {
     await this.localStorage_.copyAssetsToTempDir(reportMeta);
 
     // In offline mode, we start a local web server to test on instead of using GCS.
-    if (!isOnline) {
+    if (this.cli_.isOffline()) {
       await this.startTemporaryHttpServer_(reportMeta);
     }
 
@@ -368,7 +382,7 @@ class ReportBuilder {
    * @private
    */
   async createReportMetaProto_() {
-    const isOnline = await this.cli_.isOnline();
+    const isOnline = this.cli_.isOnline();
 
     // We only need to start up a local web server if the user is running in offline mode.
     // Otherwise, HTML files are uploaded to (and served) by GCS.
@@ -399,10 +413,10 @@ class ReportBuilder {
     const gitStatus = GitStatus.fromObject(await this.gitRepo_.getStatus());
 
     /** @type {!mdc.proto.DiffBase} */
-    const goldenDiffBase = await this.cli_.parseGoldenDiffBase();
+    const goldenDiffBase = await this.diffBaseParser_.parseGoldenDiffBase();
 
     /** @type {!mdc.proto.DiffBase} */
-    const snapshotDiffBase = await this.cli_.parseDiffBase('HEAD');
+    const snapshotDiffBase = await this.diffBaseParser_.parseDiffBase('HEAD');
 
     /** @type {!mdc.proto.GitRevision} */
     const goldenGitRevision = goldenDiffBase.git_revision;
@@ -679,6 +693,7 @@ class ReportBuilder {
       });
 
       for (const userAgent of allUserAgents) {
+        const maxRetries = this.cli_.isOnline() ? this.cli_.retries : 0;
         const userAgentAlias = userAgent.alias;
         const isScreenshotRunnable = isHtmlFileRunnable && userAgent.is_runnable;
         const expectedScreenshotImageUrl = goldenFile.getScreenshotImageUrl({htmlFilePath, userAgentAlias});
@@ -697,7 +712,7 @@ class ReportBuilder {
           actual_html_file: actualHtmlFile,
           expected_image_file: expectedImageFile,
           retry_count: 0,
-          max_retries: this.cli_.retries,
+          max_retries: maxRetries,
         }));
       }
     }
