+++ test/screenshot/infra/types/cbt-api-externs.js
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,81 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/* eslint-disable no-unused-vars */
-
-
-/*
- * Report UI
- */
-
-
-/**
- * @typedef {{
- *   checkedUserAgentCbEls: !Array<!HTMLInputElement>,
- *   uncheckedUserAgentCbEls: !Array<!HTMLInputElement>,
- *   unreviewedUserAgentCbEls: !Array<!HTMLInputElement>,
- *   changelistDict: !ReportUiChangelistDict,
- *   reviewStatusCountDict: !ReportUiReviewStatusCountDict,
- * }} ReportUiState
- */
-
-/**
- * @typedef {{
- *   changed: !ReportUiChangelistState,
- *   added: !ReportUiChangelistState,
- *   removed: !ReportUiChangelistState,
- * }} ReportUiChangelistDict
- */
-
-/**
- * @typedef {{
- *   cbEl: !HTMLInputElement,
- *   countEl: !HTMLElement,
- *   reviewStatusEl: !HTMLElement,
- *   checkedUserAgentCbEls: !Array<!HTMLInputElement>,
- *   uncheckedUserAgentCbEls: !Array<!HTMLInputElement>,
- *   reviewStatusCountDict: !ReportUiReviewStatusCountDict,
- *   pageDict: !ReportUiPageDict,
- * }} ReportUiChangelistState
- */
-
-/**
- * @typedef {!Object<string, !ReportUiPageState>} ReportUiPageDict
- */
-
-/**
- * @typedef {{
- *   cbEl: !HTMLInputElement,
- *   countEl: !HTMLElement,
- *   reviewStatusEl: !HTMLElement,
- *   checkedUserAgentCbEls: !Array<!HTMLInputElement>,
- *   uncheckedUserAgentCbEls: !Array<!HTMLInputElement>,
- *   reviewStatusCountDict: !ReportUiReviewStatusCountDict,
- * }} ReportUiPageState
- */
-
-/**
- * @typedef {!Object<string, number>} ReportUiReviewStatusCountDict
- */
-
-
-/*
- * CLI args
- */
-
-
-/**
- * @typedef {{
- *   optionNames: !Array<string>,
- *   description: string,
- *   isRequired: ?boolean,
- *   type: ?string,
- *   defaultValue: ?*,
- *   exampleValue: ?string,
- * }} CliOptionConfig
- */
-
 
 /*
  * CBT (CrossBrowserTesting.com)
@@ -234,84 +159,4 @@
  *   networks: !Array<!CbtSeleniumNetwork>,
  *   commands: !Array<!CbtSeleniumNetwork>,
  * }} CbtSeleniumInfoResponse
- */
-
-
-/*
- * Resemble.js API externs
- */
-
-
-/**
- * @typedef {{
- *   rawMisMatchPercentage: number,
- *   misMatchPercentage: string,
- *   diffBounds: !ResembleApiBoundingBox,
- *   analysisTime: number,
- *   getImageDataUrl: function(text: string): string,
- *   getBuffer: function(includeOriginal: boolean): !Buffer,
- * }} ResembleApiComparisonResult
- */
-
-/**
- * @typedef {{
- *   top: number,
- *   left: number,
- *   bottom: number,
- *   right: number,
- * }} ResembleApiBoundingBox
- */
-
-
-/*
- * ps-node API externs
- */
-
-
-/**
- * @typedef {{
- *   pid: number,
- *   ppid: number,
- *   command: string,
- *   arguments: !Array<string>,
- * }} PsNodeProcess
- */
-
-
-/*
- * Node.js API
- */
-
-
-/**
- * @typedef {{
- *   cwd: ?string,
- *   env: ?Object,
- *   argv0: ?string,
- *   stdio: ?Array<string>,
- *   detached: ?boolean,
- *   uid: ?number,
- *   gid: ?number,
- *   shell: ?boolean,
- *   windowsVerbatimArguments: ?boolean,
- *   windowsHide: ?boolean,
- * }} ChildProcessSpawnOptions
- */
-
-/**
- * @typedef {{
- *   status: number,
- *   signal: ?string,
- *   pid: number,
- * }} ChildProcessSpawnResult
- */
-
-
-/*
- * Image cropping
- */
-
-
-/**
- * @typedef {{r: number, g: number, b: number, a: number}} RGBA
  */
