+++ test/screenshot/infra/types/report-ui-types.js
@@ -1,0 +1,69 @@
+/*
+ * Copyright 2018 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Report UI
+ */
+
+
+/**
+ * @typedef {{
+ *   checkedUserAgentCbEls: !Array<!HTMLInputElement>,
+ *   uncheckedUserAgentCbEls: !Array<!HTMLInputElement>,
+ *   unreviewedUserAgentCbEls: !Array<!HTMLInputElement>,
+ *   changelistDict: !ReportUiChangelistDict,
+ *   reviewStatusCountDict: !ReportUiReviewStatusCountDict,
+ * }} ReportUiState
+ */
+
+/**
+ * @typedef {{
+ *   changed: !ReportUiChangelistState,
+ *   added: !ReportUiChangelistState,
+ *   removed: !ReportUiChangelistState,
+ * }} ReportUiChangelistDict
+ */
+
+/**
+ * @typedef {{
+ *   cbEl: !HTMLInputElement,
+ *   countEl: !HTMLElement,
+ *   reviewStatusEl: !HTMLElement,
+ *   checkedUserAgentCbEls: !Array<!HTMLInputElement>,
+ *   uncheckedUserAgentCbEls: !Array<!HTMLInputElement>,
+ *   reviewStatusCountDict: !ReportUiReviewStatusCountDict,
+ *   pageDict: !ReportUiPageDict,
+ * }} ReportUiChangelistState
+ */
+
+/**
+ * @typedef {!Object<string, !ReportUiPageState>} ReportUiPageDict
+ */
+
+/**
+ * @typedef {{
+ *   cbEl: !HTMLInputElement,
+ *   countEl: !HTMLElement,
+ *   reviewStatusEl: !HTMLElement,
+ *   checkedUserAgentCbEls: !Array<!HTMLInputElement>,
+ *   uncheckedUserAgentCbEls: !Array<!HTMLInputElement>,
+ *   reviewStatusCountDict: !ReportUiReviewStatusCountDict,
+ * }} ReportUiPageState
+ */
+
+/**
+ * @typedef {!Object<string, number>} ReportUiReviewStatusCountDict
+ */
