+++ test/screenshot/spec/mdc-drawer/classes/permanent.html
@@ -83,7 +83,7 @@
         </header>
 
         <div class="mdc-top-app-bar--fixed-adjust">
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
@@ -92,7 +92,7 @@
             occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit<br>
             anim id est laborum.
           </p>
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
@@ -101,7 +101,7 @@
             occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit<br>
             anim id est laborum.
           </p>
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
@@ -110,7 +110,7 @@
             occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit<br>
             anim id est laborum.
           </p>
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
