+++ test/screenshot/spec/mdc-drawer/classes/persistent.html
@@ -84,7 +84,7 @@
         </header>
 
         <div class="mdc-top-app-bar--fixed-adjust">
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
@@ -93,7 +93,7 @@
             occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit<br>
             anim id est laborum.
           </p>
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
@@ -102,7 +102,7 @@
             occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit<br>
             anim id est laborum.
           </p>
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
@@ -111,7 +111,7 @@
             occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit<br>
             anim id est laborum.
           </p>
-          <p>
+          <p class="test-drawer-paragraph">
             Lorem ipsum dolor sit amet, consectetur adipiscing elit,<br>
             sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut<br>
             enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut<br>
